pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentworks"
version = "1.0.0"
description = "AgentWorks - self-hosted AI agent runtime with state machine orchestration, checkpointing, and cost attribution"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    # Core data modeling - every model in the system is Pydantic v2
    "pydantic>=2.6,<3",

    # Async HTTP client - used for tool execution (Phase 3) and LLM calls (Phase 5)
    "httpx>=0.27,<1",

    # JSON Schema validation - tool input/output validation (Phase 3)
    "jsonschema>=4.21,<5",

    # Accurate token counting - Phase 6 rejected chars/4 estimates (15-30% error)
    "tiktoken>=0.6,<1",

    # API framework
    "fastapi>=0.109,<1",
    "uvicorn[standard]>=0.27,<1",

    # Redis async client - hot checkpoint store (Phase 4), response cache (Phase 5)
    "redis[hiredis]>=5.0,<6",

    # PostgreSQL async client - cold store, cost records, vector memory
    "asyncpg>=0.29,<1",

    # Configuration loading - YAML file support (Phase 8)
    "pyyaml>=6.0,<7",

    # OpenTelemetry - traces, metrics, structured logging (Phase 7)
    "opentelemetry-api>=1.22,<2",
    "opentelemetry-sdk>=1.22,<2",
    "opentelemetry-exporter-otlp>=1.22,<2",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "hypothesis>=6.98",  # Property-based testing - found 3 state machine bugs
    "ruff>=0.2",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
    "types-jsonschema>=4.21",
    "fakeredis>=2.21",   # In-memory Redis for tests
]

[tool.hatch.build.targets.wheel]
packages = ["src/agentworks"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
"src/agentworks/api.py" = ["B008"]  # FastAPI Depends()/Query() in function defaults is idiomatic

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
